+++ docs/root/intro/version_history.rst
@@ -3,6 +3,7 @@ Version history
 
 1.8.0 (Pending)
 ===============
+* config: v1 disabled by default (but still supported!)
 * http: response filters not applied to early error paths such as http_parser generated 400s.
 * ratelimit: added support for :repo:`api/envoy/service/ratelimit/v2/rls.proto`.
   Lyft's reference implementation of the `ratelimit <https://github.com/lyft/ratelimit>`_ service also supports the data-plane-api proto as of v1.1.0.